First pass at adding prefer-standard feature
This adds an event listener which watches for the view to become active. When it does it selectively enables/disables the standard and rubocop linters based on examination of some settings and the presence of a correct .yml config file.

The feature does nothing unless:

* Rubocop is installed
* Rubocop linter setting for `disable` is set to null
* Standard linter setting for `prefer_standard` is set to true

If the above are all true then we check the current project path and parent directories until we find a Standard or Rubocop config file. The first one we find tells us which linter to use. If we find none, we default to using Standard only.